refactor(react-app): 重构并大幅简化应用服务目录结构，只保留 LocalStorage 一种内置实现
- 移除 LocalStorage 外的其他实现，将单应用和多应用模式的服务拆分为 [role-mode].ts 的平铺实现，大幅简化目录结构
- 更新相关页面引用路径以适配新的服务结构
