--- packages/react-app/src/services/app/multiApp/index.ts
@@ -1,3 +1,0 @@
-import DifyAppService from './localstorage'
-
-export const appService = new DifyAppService()
