--- packages/react-app/src/services/app/multiApp/static-readonly/data.ts
@@ -1,33 +1,0 @@
-import { AppModeEnums, IDifyAppItem } from '@dify-chat/core'
-
-/**
- * 静态的应用列表，用于演示
- */
-export const staticAppList: IDifyAppItem[] = [
-	{
-		id: '0.270357011315995',
-		info: {
-			name: 'My Chatflow APP',
-			description: '我的 Chatflow 应用',
-			tags: [],
-			mode: AppModeEnums.CHATFLOW,
-		},
-		requestConfig: {
-			apiBase: 'https://api.dify.ai/v1',
-			apiKey: 'app-xxxxxxx',
-		},
-	},
-	{
-		id: '0.28936574761079314',
-		info: {
-			name: 'My Workflow APP',
-			description: '我的 Workflow 应用',
-			tags: [],
-			mode: AppModeEnums.WORKFLOW,
-		},
-		requestConfig: {
-			apiBase: 'http://127.0.0.1:5001/v1',
-			apiKey: 'app-xxxxxxx',
-		},
-	},
-]
