--- packages/react-app/src/services/app/singleApp/index.ts
@@ -1,1 +1,0 @@
-export { appConfig } from './localstorage'
